requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
boto3==1.28.17
psycopg2-binary==2.9.9
python-dotenv==1.0.1